virtual commands refactor and tests
