+++ spec/executeVirtualCommand/executeVirtualCommandSpec.ts
@@ -1,9 +1,14 @@
-import { compare, execute, executeAndReturnOutputFile, buildInputFile } from '../src'
-import { absolutize } from './testUtil'
+import { compare, execute, executeAndReturnOutputFile, buildInputFile } from '../../src'
+import { absolutize } from '../testUtil'
 
 export default describe('executeVirtualCommand', () => {
 
+  xdescribe('register new one', () => {
+    it('should allow me to register new virtual commands programmatically', async done => { done() })
+  })
+
   describe('ls', () => {
+
     it('virtual command ls found', async done => {
       const result = await execute(`
       convert rose: -rotate 55 out1.png
@@ -11,9 +16,9 @@ export default describe('executeVirtualCommand', () => {
       `)
       expect(result.exitCode).toBe(0)
       expect(result.stdout.join('\n')).toContain('out1.png')
-
       done()
     })
+
     it('virtual command ls not found', async done => {
       const result = await execute(`
       convert rose: -rotate 55 out1.png
@@ -24,6 +29,7 @@ export default describe('executeVirtualCommand', () => {
       expect(result.stderr.join('\n')).toContain('nonex.png')
       done()
     })
+
     it('virtual command ls wildcard', async done => {
       const result = await execute(`
       convert rose: -rotate 55 foo_out1.png
@@ -36,9 +42,11 @@ export default describe('executeVirtualCommand', () => {
       expect(result.stdout).toEqual(['foo_out1.png', 'foo_jo__jo_out4.png'])
       done()
     })
+
   })
 
   describe('cat', () => {
+
     it('virtual command cat found', async done => {
       const result = await execute(`
         convert logo: -format '%[pixel:p{0,0}]' info:color.txt
@@ -76,30 +84,8 @@ export default describe('executeVirtualCommand', () => {
     })
   })
 
-  describe('substitution', () => {
-    it('substitution', async done => {
-      const result = await execute(`
-      convert logo: -format '%[pixel:p{0,0}]' info:color.txt
-      convert -size 10x6 xc:skyblue  -fill \`cat color.txt\` \\
-      -draw 'point 3,2'         -scale 100x60   draw_point.gif
-    `)
-      expect(result.exitCode).toBe(0)
-      expect(await compare(result.outputFiles[1],
-        await executeAndReturnOutputFile(`convert -size 10x6 xc:skyblue -fill white -draw 'point 3,2' -scale 100x60 draw_point2.gif`))).toBe(true)
-      done()
-    })
-    it('substitution command error', async done => {
-      const result = await execute(`
-        convert -size 10x6 xc:skyblue  -fill \`cat nonex.txt\` \\
-        -draw 'point 3,2'         -scale 100x60   draw_point.gif
-    `)
-      expect(result.exitCode).not.toBe(0)
-      expect(result.stderr.join('\n')).toContain('nonex.txt')
-      done()
-    })
-  })
-
   describe('buildInputFile', () => {
+
     it('buildInputFile good', async done => {
       const result = await execute(`
       convert \`buildInputFile fn.png\` -rotate 22 out.gif
@@ -109,6 +95,7 @@ export default describe('executeVirtualCommand', () => {
       expect(await compare(result.outputFiles.slice(1, 3))).toBe(true)
       done()
     })
+
     it('buildInputFile with error', async done => {
       const result = await execute(`
          convert \`buildInputFile dontexists.png\` -rotate 22 out.gif
@@ -116,6 +103,7 @@ export default describe('executeVirtualCommand', () => {
       expect(result.exitCode).not.toBe(0)
       done()
     })
+
     it('buildInputFile from absolute url and custom name', async done => {
       const result = await execute(`
          convert \`buildInputFile '${absolutize('fn.png')}' custom.png\` -rotate 22 out.gif
@@ -124,69 +112,34 @@ export default describe('executeVirtualCommand', () => {
       expect(result.exitCode).toBe(0)
       done()
     })
+
   })
 
   describe('uniqueName', () => {
+
     it('uniqueName', async done => {
+      const result = await execute(`
+      convert rose: -rotate 22 foo\`uniqueName\`.gif
+      convert rose: -rotate 22 \`uniqueName\`.gif
+    `)
+      expect(result.exitCode).toBe(0)
+      expect(result.outputFiles[0].name.startsWith('foo')).toBe(true)
+      expect(result.outputFiles[0].name.endsWith('.gif')).toBe(true)
+      done()
+    })
+
+    it('uniqueName gives unique names', async done => {
       const result = await execute(`
       convert rose: -rotate 22 \`uniqueName\`.gif
     `)
       expect(result.exitCode).toBe(0)
-      expect(result.outputFiles.length).toBe(1)
+      const result2 = await execute(`
+      convert rose: -rotate 22 \`uniqueName\`.gif
+    `)
+      expect(result2.exitCode).toBe(0)
+      expect(result.outputFiles[0].name).not.toBe(result2.outputFiles[0].name)
       done()
     })
   })
 
-  describe('variable decl', () => {
-    it('variable decl good', async done => {
-      const result = await execute(`
-      color='white'
-      convert -size 10x6 xc:skyblue  -fill $color \\
-      -draw 'point 3,2' -scale 100x60   draw_point.gif
-    `)
-      expect(result.exitCode).toBe(0)
-      expect(await compare(result.outputFiles[0],
-        await executeAndReturnOutputFile(`convert -size 10x6 xc:skyblue -fill white -draw 'point 3,2' -scale 100x60 draw_point2.gif`))).toBe(true)
-      done()
-    })
-    it('variable decl with substitution and other reference in the same command', async done => {
-      const result = await execute(`
-      img1='fn.png'
-      img2='foo.png'
-      convert \`buildInputFile $img1\` -resize 45% $img2
-    `)
-      expect(result.exitCode).toBe(0)
-      expect(result.outputFiles[1].name).toBe('foo.png')
-      const result2 = await execute([await buildInputFile('fn.png')], `convert fn.png -resize 45% output.png`)
-      expect(await compare(result.outputFiles[1], result2.outputFiles[0])).toBe(true)
-      done()
-    })
-
-    xit('variable decl from substitution ouput', async done => { // TODO: this is failing - probably we need to separate variable declaration in two plugins - variable-declaration and variable references. and put variable declarations AFTER substitution
-      const result = await execute(`
-         var1='\`identify rose:\`'
-    `)
-      expect(result.exitCode).toBe(0)
-      debugger
-      done()
-    })
-
-    // it('variable decl with substitution with error', async done => {
-    //   const result = await execute(`
-    //      var1='\`cat noexist.txt\`'
-    // `)
-    //   expect(result.exitCode).not.toBe(0)
-    //   debugger
-    //   done()
-    // })
-    // it('variable decl wrong - used not existent', async done => {
-    //     const result = await execute(`
-    //     color=white
-
-    // `)
-    //     expect(result.exitCode).not.toBe(0)
-    //     expect(result.stderr.join('\n')).toContain('nonex.txt')
-    //     done()
-    //   })
-  })
 })
