+++ spec/index.ts
@@ -5,7 +5,9 @@ export * from './executeSpec'
 export * from './formatSpec'
 export * from './imageHomeSpec'
 export * from './executionContextSpec'
-export * from './executeVirtualCommandSpec'
+export * from './executeVirtualCommand/executeVirtualCommandSpec'
+export * from './executeVirtualCommand/substitutionSpec'
+export * from './executeVirtualCommand/variableSpec'
 export * from './testAssetsSpec'
 export * from './util/cliSpec'
 export * from './util/htmlSpec'
